added null check to web worker on unmount

+++ frontend/src/components/ModalIdle.vue
@@ -107,7 +107,9 @@ export default {
     },
   },
   beforeUnmount() {
-    this.countdownWorker.terminate();
+    if (this.countdownWorker) {
+      this.countdownWorker.terminate();
+    }
   },
 };
 </script>
